Trait beacon::Beacon

pub trait Beacon where
    Self: Sized, {
#[must_use]    fn verify_entry<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        curr: &'life1 BeaconEntry, 
        prev: &'life2 BeaconEntry
    ) -> Pin<Box<dyn Future<Output = Result<bool, Box<dyn Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn entry<'life0, 'async_trait>(
        &'life0 self, 
        round: u64
    ) -> Pin<Box<dyn Future<Output = Result<BeaconEntry, Box<dyn Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn max_beacon_round_for_epoch(&self, fil_epoch: ChainEpoch) -> u64;
}

Required methods

#[must_use]fn verify_entry<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
curr: &'life1 BeaconEntry,
prev: &'life2 BeaconEntry
) -> Pin<Box<dyn Future<Output = Result<bool, Box<dyn Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Verify a new beacon entry against the most recent one before it.

#[must_use]fn entry<'life0, 'async_trait>(
&'life0 self,
round: u64
) -> Pin<Box<dyn Future<Output = Result<BeaconEntry, Box<dyn Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a BeaconEntry given a round. It fetches the BeaconEntry from a Drand node over GRPC In the future, we will cache values, and support streaming.

fn max_beacon_round_for_epoch(&self, fil_epoch: ChainEpoch) -> u64

Loading content...